moving screen setup in hypervisor
