+++ .gitignore
@@ -2,4 +2,4 @@
 *.xml
 .idea/
 main
-vsicript
+viscript
