+++ hypervisor/hypervisor.go
@@ -5,7 +5,68 @@ import (
 	"github.com/corpusc/viscript/gfx"
 	"github.com/corpusc/viscript/script"
 	"github.com/go-gl/gl/v2.1/gl"
+	"log"
+	"github.com/go-gl/glfw/v3.2/glfw"
+	"github.com/corpusc/viscript/app"
+	"github.com/corpusc/viscript/msg"
+	"runtime"
 )
+
+
+
+
+func init() {
+	fmt.Println("main.init()")
+	gfx.MakeHighlyVisibleLogHeader(app.Name, 15)
+	// GLFW event handling must run on the main OS thread
+	// See documentation for functions that are only allowed to be called from the main thread.
+	runtime.LockOSThread()
+}
+
+
+func ScreenSetup(){
+
+	fmt.Printf("Start\n")
+
+	if err := glfw.Init(); err != nil {
+		log.Fatalln("failed to initialize glfw:", err)
+	}
+
+	defer glfw.Terminate()
+
+	glfw.WindowHint(glfw.Resizable, glfw.True)
+	glfw.WindowHint(glfw.ContextVersionMajor, 2)
+	glfw.WindowHint(glfw.ContextVersionMinor, 1)
+	window, err := glfw.CreateWindow(int(gfx.CurrAppWidth), int(gfx.CurrAppHeight), app.Name, nil, nil)
+
+	if err != nil {
+		panic(err)
+	}
+
+	window.MakeContextCurrent()
+
+	if err := gl.Init(); err != nil {
+		panic(err)
+	}
+	Texture = NewTexture("Bisasam_24x24_Shadowed.png")
+
+	defer gl.DeleteTextures(1, &Texture)
+
+	InitRenderer()
+
+	InitInputEvents(window)
+
+	for !window.ShouldClose() {
+		msg.MonitorEvents(Events)
+		glfw.PollEvents()
+		DrawScene()
+		//drawScene()
+		window.SwapBuffers()
+	}
+
+
+
+}
 
 func InitRenderer() {
 	fmt.Println("InitRenderer()")
